lustre: reintroduce ATTR_ATTR_FLAG
In Linux kernels >= 4.19, the ATTR_ATTR_FLAG flag was removed because
nobody was using it.  But Lustre does. So define it if needed. It is
safe to do so, since it's not been replaced with a different flag with
the same value.

Linux commit: 4cdfffc8722e ("vfs: discard ATTR_ATTR_FLAG")

Signed-off-by: Frank van der Linden <fllinden@amazon.com>