lustre: adapt for totalram_pages change
totalram_pages was changed to a function in Linux 5.0. Adapt for this
change by hiding it behind a TOTALRAM_PAGES define, which is defined
differently depending on the HAVE_TOTALRAM_PAGES_FUNC config. Add the
right include files to the files hat need to check this configuration.

Linux commit: ca79b0c211af ("mm: convert totalram_pages and totalhigh_pages variables to atomic")

Signed-off-by: Frank van der Linden <fllinden@amazon.com>