Add Amazon EFA driver version 1.11.1
Add the EFA driver to the 5.10 tree. This squashes three cherry-picked
commits from amazon-5.4.y/master:
b85486ce "Add Amazon EFA driver version 1.4"
499ab763 "drivers/amazon: efa: update to 1.9.0"
2ba3701e "drivers/amazon: efa: update to 1.10.1"

Then the driver is updated to 1.11.1. This is done because an update to
config.h is necessary to solve a redefinition of structs in
include/rdma/ib_verbs.h, so we may as well update while we're here.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>