xen-blkfront: Delay flush till queue lock dropped
In the Amazon Linux 4.14.y kernel the xen-blkfront driver can encounter
a deadlock condition when handling certain errors from the backend:

[   60.870492] BUG: spinlock recursion on CPU#1, swapper/1/0
[   60.874250]  lock: 0xffff8881e096ac00, .magic: dead4ead, .owner: swapper/1/0, .owner_cpu: 1
[   60.879879] CPU: 1 PID: 0 Comm: swapper/1 Not tainted 4.14.173-138.231.amzn2.x86_64 #1
[   60.885249] Hardware name: Xen HVM domU, BIOS 4.2.amazon 08/24/2006
[   60.889623] Call Trace:
[   60.891664]  <IRQ>
[   60.893420]  dump_stack+0x66/0x85
[   60.895987]  do_raw_spin_lock+0x8d/0xc0
[   60.898884]  blk_queue_write_cache+0x1d/0x140
[   60.902041]  xlvbd_flush+0x26/0xa0 [xen_blkfront]

This is caused by the blkfront interrupt handler updating the write
cache mode; blk_queue_write_cache() tries to lock the queue lock, which
has already been locked by the handler. The blk_queue_write_cache() call
can't be delayed to the end of the handler when it drops the lock since
it locks the queue lock with spin_lock_irq(), and accidentally enables
interrupts before the handler has returned. Instead add a work_struct to
blkfront_info and defer the write cache update until after the handler
has returned.

This bug was introduced as part of the "xen-blkfront: resurrect
request-based mode" patch. While the 4.14.y kernel will move away from
this and use blk-mq instead soon, this is a smaller-impact fix for
current users.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>
Reviewed-by: Frank van der Linden <fllinden@amazon.com>
Reviewed-by: Andy Strohman <astroh@amazon.com>
Reviewed-by: Balbir Singh <sblbir@amazon.com>