lustre: seperate debugfs and procfs handling
This first part is a partial backport of:

    LU-13344 all: Separate debugfs and procfs handling

    Linux 5.6 introduces proc_ops with v5.5-8862-gd56c0d45f0e2
    proc: decouple proc from VFS with "struct proc_ops"

    Separate debugfs usage and procfs usage to prepare for the divergence
    of debugfs using file_operations and procfs using proc_ops

    HPE-bug-id: LUS-8589
    Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
    Change-Id: I1746e563b55a9e89f90ac01843c304fe6b690d8b
    Reviewed-on: https://review.whamcloud.com/37834
    Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
    Reviewed-by: Neil Brown <neilb@suse.de>
    Reviewed-by: James Simmons <jsimmons@infradead.org>
    Tested-by: jenkins <devops@whamcloud.com>
    Tested-by: Maloo <maloo@whamcloud.com>
    Reviewed-by: Oleg Drokin <green@whamcloud.com>

The backport is partial, because the debugfs code isn't actually used
in the Lustre 2.10 branch. But this is needed for the file_operations ->
proc_ops change in Linux 5.6, which hasn't been completed yet in upstream
Lustre. Get ready for that change by also changing file_operations
to proc_ops in lprocfs_var.

The second part os converting all remaining struct file_operations use
for proc ops to struct proc_ops.

Signed-off-by: Frank van der Linden <fllinden@amazon.com>