arm64: entry: Add annotation when switching to/from the irq stack
Handling of interrupts is done on a separate stack. The stack switch
can confuse objtool.

When starting interrupt handling, the only valid assumption is that the
stack pointer points to a valid address. After handling an interrupt,
all that is known is that the stack pointer points to a pt_regs frame.

Signed-off-by: Julien Thierry <jthierry@redhat.com>