mm/damon: Introduce arbitrary target type
A problem of region-based monitoring is additional memory space overhead
for the regions metadata.  For example, suppose we want to monitor if
each page is acceessed or not rather than getting the access frequency
of each memory region.  Without DAMON, we can do that by periodically
resetting and reading the PG_Idle flags and/or the PTE Accessed bits of
the pages.  We can also use DAMON for that by setting the minimal number
of regions and maximum number of regions as 'the monitoring target
region size / page size', and setting the sampling interval same to the
aggregation interval.

However, the metadata for the region abstraction (start address, end
address, and observed accesses counter of each region) is only a huge
burden in the case, compared to the without-DAMON strategy.  We could
implement another monitoring primitive that doesn't use the 'struct
damon_region' abstraction.  However, because the damon region
abstraction is strongly coupled with the core logic, it would be not
straightforward.

To support such cases, this commit makes DAMON to support another type
of monitoring targets that use user-defined arbitrary monitoring target
regions instead of the 'struct damon_region'.