Revert: crypto: jitterentropy - change back to module_init()
The original patch assumes that self tests are not run jitterentropy.
That is not true on this kernel. When self tests are called on jitterentropy
the kernel panics because the module is not yet loaded. Reverting this patch
causes jitternetropy to loaded with subsys_initcall(). This, in turn, causes
the module to be loaded in time for the self tests.