Revert "gup: document and work around "COW can break either way" issue"
This reverts commit 407faed92b4a4e2ad900d61ea3831dd597640f29.

The above commit was introduced to address CVE-2020-29374.

It caused many bugs and was eventually reverted by:
1a0cf26323c8 mm/ksm: Remove reuse_ksm_page()
a308c71bf1e6 mm/gup: Remove enfornced COW mechanism
798a6b87ecd7 mm: Add PGREUSE counter
Once a different fixed for the CVE was merged:
09854ba94c6a mm: do_wp_page() simplification

This alternative fix also had bugs, some of which remain unaddressed
upstream. Additionally the vulnerability the CVE addresses still exists
with the above patch when huge pages are used.

For now this CVE is considered won't fix for Amazon Linux so revert
the original commit as it introduces known regressions.