xen-blkfront: fix a race of 'persistent_grants' feature
commit 74a852479c68e4efb3865d5436fd69ec4f819f96 upstream.

Xen blkfront module allows the backend to know if it supports persistent
grants feature in 'talk_to_blkback()', and then checks if the backend
supports the feature or not in 'blkfront_gather_backend_features()'.  It
uses the persistent grants feature only if both ends support the
feature.  The backend also has similar mechanism.

A parameter for disabling persistent grants feature from blkfront module
called 'persistent_grants', which is introduced by our downstream commit
e2e1bb43085f ("xen-blkfront: add 'persistent_grants' parameter"), is
read twice in the above-mentioned two functions.  If user changes the
parameter between the executions of the two functions, this could result
in asymmetric use of the feature[2,3].  Detailed sequence is,

0. The backend supports persistent grants feature.
1. The user sets 'persistent_grants' parameter of blkfront as 'false'
2. 'talk_to_blkback()' is called, says backend that it doesn't support
   the feature.
3. So the backend will not use the feature.
4. The user sets 'persistent_grants' parameter of blkfront as 'true'.
5. 'blkfront_gather_backend_features()' shows the parameter 'true', and
   the backend supports the feature.
6. 'blkfront' will use the feature while the backend doesn't.

Upstream commit 74a852479c68 ("xen-blkfront: add a parameter for
disabling of persistent grants") introduced another Xen blkfront
parameter called 'feature_persistent', which doesn very same work.  But,
it fixed the issue by caching the parameter once and reusing it.

This commit fixes the race of the downstream commit by replacing its
internal mechanism with that of the upstream commit.  Note that it keeps
the parameter name same so that this doesn't introduce any behavioral
change.

[1] https://lore.kernel.org/linux-block/fdbaf955-0b92-d356-2792-21b27ea1087d@suse.com/
[2] https://lore.kernel.org/linux-block/20200922111259.GJ19254@Air-de-Roger/
[3] https://lore.kernel.org/linux-block/fdbaf955-0b92-d356-2792-21b27ea1087d@suse.com/

Fixes: e2e1bb43085f ("xen-blkfront: add 'persistent_grants' parameter")
Signed-off-by: SeongJae Park <sjpark@amazon.com>