Revert "PCI/MSI: Let core code free MSI descriptors"
This reverts commit 9fb9eb4b59acc607e978288c96ac7efa917153d4.

This commit causes issues with hibernation on Xen guests (null-pointer
exception on freeze), most likey due to the Xen driver trying to work on
interrupts after they've been freed.

While the downstream Xen patches are updated to work properly, revert
this change. This is modified to still use the new accessors introduced
in the same series, and just call out to msi_free_msi_descs() which does
exactly the same work.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>