Revert "selinux: runtime disable is deprecated, add some ssleep() discomfort"
This reverts commit 43b666622c60bc001f2f8a19f5f97946ff53a5cc.

Let's keep selinux run-time disable behavior the same we have
in 5.15, since AL2 user-space expects to be able to disable
selinux at run-time.

Note that the change being reverted is only in effect when
CONFIG_SECURITY_SELINUX_DISABLE=y. Which is not the case
for AL2023 but is the case for AL2.

No discomfort for AL2 users ;)

Signed-off-by: Luiz Capitulino <luizcap@amazon.com>