Revert "perf/x86/amd/core: Fix overflow reset on hotplug"
There's unchecked MSR access error to MSR_AMD64_PERF_CNTR_GLOBAL_STATUS_CLR
in all zen4 virtualized instances during amd_pmu_cpu_starting.
This commit clearer LBR freeze bit of PerfCntrGLobalStatus register
and moved the reset sequence in amd_pmu_cpu start and dead but reverting
either change can fade the MSR access error away.

It could be possible that hypervisor does not allow the write to this
register and this commit made reset in effect. To prevent error
call traces in the dmesg and unblock the release, we need to revert this
commit for now.

This reverts commit e485a69d9b4433ad48673a1d31c705b09b3548c0.

Signed-off-by: Shaoying Xu <shaoyi@amazon.com>